Removed repo count
Removed because this doesn't reflect the number of repos a user has on
styleci. This only shows the number of repos we're using this user's
access_token on to contact github. Even if it did reflect the true
number of repos on the system, it's floored because we can enable a repo
in the other section of the page, and flick back to this and the number
hasn't moved.

// cc @jbrooksuk