Fixed a bug causing fetching collaborators to fail
The bug was caused because the act of decorating a model actually sets
the relations on it to the presented ones before wrapping it and
returning it. This means further calls to the user model, inside the
user repository were getting the decorated models instead of the
originals. I wonder if we should change autopresenter to clone objects
first, because they behaviour feels quite unintuative to me. // cc
@joecohens